requirements.txt
asgiref==3.9.1
attrs==25.4.0
beautifulsoup4==4.14.2
Brotli==1.1.0
click==8.3.0
dj-database-url==3.0.1
Django==4.2.25
django-bootstrap-v5==1.0.11
django-phonenumber-field==8.1.0
django-schema-viewer==0.5.3
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
phonenumbers==9.0.13
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.37.0
whitenoise==6.11.0
